fix: correctly infer the type of the function passed to derived
Before this commit, the type of the function passed to derived was not
correctly inferred by typescript in the async case. By putting the async
case first in the list of overloads of the derived function, it looks like
it fixes the issue, and the type is still correctly inferred in the sync
case too.