chore: directly use the index.d.ts file generated by tsc (#73)
The file produced by the api-extractor is buggy because it does not contain the following needed declaration:

```ts
declare global {
    interface SymbolConstructor {
        readonly observable: symbol;
    }
}
```

Without this, the InteropObservable interface no longer means anything, and type inference fails on all derived calls.

This commit simply uses the index.d.ts file generated by tsc instead of using the result from the api-generator. In this file, the global Symbol.observable declaration is not missing and types work fine again.